Split Cell from Gui added static methods for Cell to access Gui moved coordTable to Move module
