+++ ProjectCode/test/java/NineMensMorris/GuiTest.java
@@ -1,0 +1,14 @@
+package NineMensMorris;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class GuiTest {
+
+    @org.junit.jupiter.api.BeforeEach
+    void setUp() {
+    }
+
+    @org.junit.jupiter.api.AfterEach
+    void tearDown() {
+    }
+}
